+++ packages/dataviews/src/components/dataform-layouts/card/index.tsx
@@ -180,7 +180,6 @@ export default function FormCardField< Item >( {
 		() => ( {
 			layout: DEFAULT_LAYOUT as NormalizedLayout,
 			fields: field.children ?? [],
-			labelMode: 'showRequired',
 		} ),
 		[ field ]
 	);