+++ packages/dataviews/src/components/dataform-layouts/panel/dropdown.tsx
@@ -90,7 +90,6 @@ function PanelDropdown< Item >( {
 				? field.children
 				: // If not explicit children return the field id itself.
 				  [ { id: field.id, layout: DEFAULT_LAYOUT } ],
-			labelMode: 'showRequired',
 		} ),
 		[ field ]
 	);