+++ packages/dataviews/src/components/dataform-layouts/panel/modal.tsx
@@ -59,7 +59,6 @@ function ModalContent< Item >( {
 				? field.children
 				: // If not explicit children return the field id itself.
 				  [ { id: field.id, layout: DEFAULT_LAYOUT } ],
-			labelMode: 'showRequired',
 		} ),
 		[ field ]
 	);