+++ packages/dataviews/src/components/dataform-layouts/regular/index.tsx
@@ -56,7 +56,6 @@ export default function FormRegularField< Item >( {
 		() => ( {
 			layout: DEFAULT_LAYOUT,
 			fields: !! field.children ? field.children : [],
-			labelMode: 'showRequired',
 		} ),
 		[ field ]
 	);