+++ packages/dataviews/src/components/dataform-layouts/row/index.tsx
@@ -51,7 +51,6 @@ export default function FormRowField< Item >( {
 		const form: NormalizedForm = {
 			layout: DEFAULT_LAYOUT as NormalizedLayout,
 			fields: field.children,
-			labelMode: 'showRequired',
 		};
 
 		return (